requirements.txt
kucoin-python~=1.0.14
numpy~=1.24.0
openai~=1.23.2
pandas~=2.0.0
peewee~=3.16.3
six~=1.16.0
python-dotenv~=1.0.0
pyopenssl~=23.2.0
beautifulsoup4~=4.12.2
gspread~=5.12.0
feedparser~=6.0.11
pylint~=3.0.3
dspy-ai~=2.4.9

# --- testing ---
pytest~=8.0.1
pytest-mock~=3.14.0
